# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ECOnexão API
  description: API de domínio do aplicativo ECOnexão para turismo sustentável e rotas
    ecológicas.
  version: 1.0.0
paths:
...
  /api/v1/admin/territory/regions:
    get:
      tags:
      - Admin Territorial
      summary: Listar regiões administrativas
      operationId: list_regions_api_v1_admin_territory_regions_get
      security:
      - HTTPBearer: []
      parameters:
      - name: include_inactive
        in: query
        required: false
        schema:
          type: boolean
          description: Incluir regiões inativas
          default: true
          title: Include Inactive
        description: Incluir regiões inativas
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRegionListEnvelope'
        '401':
          description: JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Admin Territorial
      summary: Criar região administrativa
      operationId: create_region_api_v1_admin_territory_regions_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRegionCreateSchema'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRegionEnvelope'
        '401':
          description: JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Slug de região já existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/territory/regions/{region_id}:
    get:
      tags:
      - Admin Territorial
      summary: Obter detalhe administrativo de uma região
      operationId: get_region_api_v1_admin_territory_regions__region_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: region_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Region Id
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRegionEnvelope'
        '401':
          description: JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Região não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Admin Territorial
      summary: Atualizar região administrativa
      operationId: update_region_api_v1_admin_territory_regions__region_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: region_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Region Id
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRegionUpdateSchema'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRegionEnvelope'
        '401':
          description: JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Região não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/territory/routes:
    get:
      tags:
      - Admin Territorial
      summary: Listar rotas administrativas
      operationId: list_routes_api_v1_admin_territory_routes_get
      security:
      - HTTPBearer: []
      parameters:
      - name: region_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Filtrar por ID de região
          title: Region Id
        description: Filtrar por ID de região
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filtrar por status
          title: Status
        description: Filtrar por status
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Busca textual por título ou cidade
          title: Q
        description: Busca textual por título ou cidade
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRouteListEnvelope'
        '401':
          description: JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Admin Territorial
      summary: Criar rota administrativa
      operationId: create_route_api_v1_admin_territory_routes_post
      security:
      - HTTPBearer: []
      parameters:
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRouteCreateSchema'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRouteEnvelope'
        '401':
          description: JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Slug de rota já existe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/admin/territory/routes/{route_id}:
    get:
      tags:
      - Admin Territorial
      summary: Obter detalhe administrativo de uma rota
      operationId: get_route_api_v1_admin_territory_routes__route_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: route_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Route Id
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRouteEnvelope'
        '401':
          description: JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rota não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - Admin Territorial
      summary: Atualizar rota administrativa
      operationId: update_route_api_v1_admin_territory_routes__route_id__put
      security:
      - HTTPBearer: []
      parameters:
      - name: route_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Route Id
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRouteUpdateSchema'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRouteEnvelope'
        '401':
          description: JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rota não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflito de concorrência ou slug.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Admin Territorial
      summary: Arquivar rota administrativa
      operationId: archive_route_api_v1_admin_territory_routes__route_id__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: route_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Route Id
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRouteEnvelope'
        '401':
          description: JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Rota não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/admin/territory/routes/{route_id}/origins/{origin_id}/geometries:
    post:
      tags:
      - Admin Territorial
      summary: Criar geometria espacial de rota
      operationId: create_geometry_api_v1_admin_territory_routes__route_id__origins__origin_id__geometries_post
      security:
      - HTTPBearer: []
      parameters:
      - name: route_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Route Id
      - name: origin_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Origin Id
      - name: X-Region-ID
        in: header
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: Escopo regional explícito da operação editorial.
          title: X-Region-Id
        description: Escopo regional explícito da operação editorial.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRouteGeometryCreateSchema'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRouteGeometryEnvelope'
        '401':
          description: JWT ausente ou inválido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Permissão negada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Origem ou rota não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Geometria já existe para este provedor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Geometria espacial malformada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/routes/{route_id}:
    get:
      tags:
      - Territorial - Routes
      summary: Detalhes de uma rota
      description: Retorna informações completas e origens de uma rota específica.
      operationId: get_route_detail_api_v1_routes__route_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: route_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Route Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDetailEnvelope'
        '404':
          description: Rota não encontrada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/routes/{route_id}/origins:
    get:
      tags:
      - Territorial - Routes
      summary: Origens de uma rota
      description: Retorna os pontos de partida registrados para a rota especificada.
      operationId: get_route_origins_api_v1_routes__route_id__origins_get
      parameters:
      - name: route_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Route Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteOriginListEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AdminRegionEnvelope:
      properties:
        data:
          $ref: '#/components/schemas/AdminRegionSchema'
      type: object
      required:
      - data
      title: AdminRegionEnvelope
    AdminRegionListEnvelope:
      properties:
        data:
          items:
            $ref: '#/components/schemas/AdminRegionSchema'
          type: array
          title: Data
      type: object
      required:
      - data
      title: AdminRegionListEnvelope
...
    AdminRouteListEnvelope:
      properties:
        data:
          items:
            $ref: '#/components/schemas/AdminRouteSchema'
          type: array
          title: Data
        meta:
          $ref: '#/components/schemas/PaginationMeta'
      type: object
      required:
      - data
      - meta
      title: AdminRouteListEnvelope
...
    ErrorResponse:
      properties:
        error:
          $ref: '#/components/schemas/ErrorDetail'
        request_id:
          type: string
          title: Request Id
          example: req_123456789
      type: object
      required:
      - error
      - request_id
      title: ErrorResponse
      description: Standard error response envelope.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer